Big Refactoring domain: left entities with Hibernate mapping, moved to new directory
