+++ myBrikProject/src/main/java/by/mybrik/domain/Goods.java
@@ -1,54 +1,65 @@
 package by.mybrik.domain;
 
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import org.apache.commons.lang3.builder.ToStringBuilder;
-import org.apache.commons.lang3.builder.ToStringStyle;
+import by.mybrik.domain.Gender;
+import lombok.Data;
 
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.EnumType;
+import javax.persistence.Enumerated;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
 import java.sql.Timestamp;
 
-@Setter
-@Getter
-@Builder
-@EqualsAndHashCode
-@NoArgsConstructor
-@AllArgsConstructor
+@Data
+@Entity
+@Table(name ="m_goods" )
 public class Goods {
 
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
+    @Column(name = "order_code")
     private String orderCode;
 
+    @Column
     private String name;
 
+    @Column
     private  String photo;
 
+    @Column
+    @Enumerated(EnumType.STRING)
     private Gender gender;
 
+    @Column
     private String size;
 
+    @Column
     private String color;
 
+    @Column
     private String description;
 
+    @Column(name = "is_deleted")
     private boolean isDeleted;
 
+    @Column
     private Double price;
 
+    @Column
     private int quantity;
 
+    @Column
     private String category;
 
+    @Column
     private Timestamp created = new Timestamp(System.currentTimeMillis());
 
+    @Column
     private Timestamp changed = new Timestamp(System.currentTimeMillis());
 
-    @Override
-    public String toString() {
-        return ToStringBuilder.reflectionToString(this, ToStringStyle.JSON_STYLE);
-    }
 }
