--- myBrikProject/src/main/java/by/mybrik/domain/TextileProductType.java
@@ -1,31 +1,0 @@
-package by.mybrik.domain;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import org.apache.commons.lang3.builder.ToStringBuilder;
-import org.apache.commons.lang3.builder.ToStringStyle;
-
-@Setter
-@Getter
-@Builder
-@EqualsAndHashCode
-@NoArgsConstructor
-@AllArgsConstructor
-public class TextileProductType {
-
-    private Long id;
-
-    private Long textileId;
-
-    private Long productTypeId;
-
-    @Override
-    public String toString() {
-        return ToStringBuilder.reflectionToString(this, ToStringStyle.JSON_STYLE);
-    }
-
-}
