+++ myBrikProject/src/main/java/by/mybrik/domain/Users.java
@@ -1,50 +1,59 @@
 package by.mybrik.domain;
 
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import org.apache.commons.lang3.builder.ToStringBuilder;
-import org.apache.commons.lang3.builder.ToStringStyle;
+import by.mybrik.domain.Gender;
+import lombok.Data;
 
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.EnumType;
+import javax.persistence.Enumerated;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
 import java.sql.Timestamp;
 
-@Setter
-@Getter
-@Builder
-@EqualsAndHashCode
-@NoArgsConstructor
-@AllArgsConstructor
+@Data
+@Entity
+@Table(name = "m_users")
 public class Users {
 
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
+    @Column
     private String name;
 
+    @Column
     private String surName;
 
+    @Column
     private String login;
 
+    @Column
     private String password;
 
+    @Column
     private String email;
 
+    @Column
+    @Enumerated(EnumType.STRING)
     private Gender gender;
 
+    @Column
     private Timestamp created = new Timestamp(System.currentTimeMillis());
 
+    @Column
     private Timestamp changed = new Timestamp(System.currentTimeMillis());
 
+    @Column
     private String phone;
 
+    @Column
     private String address;
 
+    @Column(name = "is_deleted")
     private boolean isDeleted;
 
-    @Override
-    public String toString() {
-        return ToStringBuilder.reflectionToString(this, ToStringStyle.JSON_STYLE);
-    }
 }
