--- myBrikProject/src/main/java/by/mybrik/domain/entities/IndividualOrder.java
@@ -1,48 +1,0 @@
-package by.mybrik.domain.entities;
-
-import lombok.Data;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Table;
-import java.sql.Timestamp;
-
-@Data
-@Entity
-@Table(name = "m_individual_order")
-public class IndividualOrder {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(name = "user_id")
-    private Long userId;
-
-    @Column(name = "textile_id")
-    private Long textileId;
-
-    @Column(name = "product_type_id")
-    private Long productTypeId;
-
-    @Column(name = "price_id")
-    private Long priceId;
-
-    @Column
-    private int quantity;
-
-    @Column(name = "total_price")
-    private Double totalPrice;
-
-    @Column(name = "order_status")
-    private String orderStatus;
-
-    @Column
-    private Timestamp created = new Timestamp(System.currentTimeMillis());
-
-    @Column
-    private Timestamp changed = new Timestamp(System.currentTimeMillis());
-}
