+++ myBrikProject/src/main/java/by/mybrik/repository/GoodsRepository.java
@@ -1,6 +1,6 @@
 package by.mybrik.repository;
 
-import by.mybrik.domain.entities.Goods;
+import by.mybrik.domain.Goods;
 
 public interface GoodsRepository extends CrudRepository<Long, Goods> {
