+++ myBrikProject/src/main/java/by/mybrik/repository/ProductTypeRepository.java
@@ -1,6 +1,6 @@
 package by.mybrik.repository;
 
-import by.mybrik.domain.entities.ProductType;
+import by.mybrik.domain.ProductType;
 
 public interface ProductTypeRepository extends CrudRepository<Long, ProductType> {
 }
