+++ myBrikProject/src/main/java/by/mybrik/repository/StandardOrderRepository.java
@@ -1,6 +1,6 @@
 package by.mybrik.repository;
 
-import by.mybrik.domain.entities.StandardOrder;
+import by.mybrik.domain.StandardOrder;
 
 public interface StandardOrderRepository extends CrudRepository<Long, StandardOrder> {
 }
