+++ myBrikProject/src/main/java/by/mybrik/repository/UsersRepository.java
@@ -1,6 +1,6 @@
 package by.mybrik.repository;
 
-import by.mybrik.domain.entities.Users;
+import by.mybrik.domain.Users;
 
 public interface UsersRepository extends CrudRepository<Long, Users> {
 }
