+++ myBrikProject/src/main/java/by/mybrik/repository/implementation/IndividualOrderImpl.java
@@ -1,6 +1,6 @@
 package by.mybrik.repository.implementation;
 
-import by.mybrik.domain.entities.IndividualOrder;
+import by.mybrik.domain.IndividualOrder;
 import by.mybrik.repository.IndividualOrderRepository;
 import lombok.RequiredArgsConstructor;
 import org.hibernate.Session;
