+++ myBrikProject/src/main/java/by/mybrik/service/TextileService.java
@@ -1,7 +1,7 @@
 package by.mybrik.service;
 
 
-import by.mybrik.domain.entities.Textile;
+import by.mybrik.domain.Textile;
 
 import java.util.List;
 import java.util.Optional;
